# lib/api-zod/src/generated/types/businessHourEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessHourEntryDayOfWeek } from './businessHourEntryDayOfWeek';

export interface BusinessHourEntry {
  id: string;
  businessId: string;
  dayOfWeek: BusinessHourEntryDayOfWeek;
  openTime: string;
  closeTime: string;
  isClosed: boolean;
}